refactor: move transform validators to utility header
